run-workload.sh: changed the way the output directory is used
The fio daemons on the clients need the same output directory if this is
part of the output file name. If it doesn't exist the fio task simply hangs.

This change uses the default location for the fio daemon, and then copies
the resulting output files to the required location after the run completes