genchart: updated to provide a text summary
At the end of the analysis a text based summary is now provided for
a quick overview of the json statistics. This is especially useful when
the fio run is just a single test run of 'n' clients - in this scenario a chart
would only have a single data point for each category.